Add accept_follow_request and reject_follow_request
The accept_follow_request method works by establishing the relationship
between the two users and updating the status of the request itself.
It is called from a form from the users/index partial.

The partial now also displays received follow requests and followers.

Next to a request will be the two buttons that allow you to accept or
delete a request, calling their respective methods.

The reject_follow_request simply changes the status of the request to
'rejected' without changing the relationship at all.