Add unfollow method
This is made possible by three main components:

The unfollow_button helper that generates a form with the user's id
and a button that links to the #unfollow method of the UsersController.

The #unfollow method itself that removes the follower-followee relationship
between the current_user and the followed user while deleting the previously
accepted request.

The delete_request method in the FollowRequest model that finds and deletes
the request using the ids of the two users. I chose to delete it because I
have no interest in tracking request histories.