Add profile pictures
Using ActiveStorage, profile pictures can now be uploaded from the profile
view of the current_user. To do this, I had to add an update method to
the UsersController, a #profile_picture_form helper in users_helper
to generate the form, an additional relationship to the User model and the
edit/update routes.

I have made it so that the users index shows a minimized version of the
profile picture of every user and the profile page of a user shows a larger
version of that profile picture.