Add ability to create posts
A new post is initiated in the index method of posts because that is my
home page. Needed a create method to create posts.

I set up the relationship between a user that has_many posts and a post
that belong_to one author, adding a user_id foreign_key to my posts table.

A _form partial was made and rendered on the index page to allow the
current_user to create posts.

For now, all posts will be shuffled and made visible to the current_user,
while the individual posts of a user will all appear on their profile page.

I have also realized that I was writing image_tags to display profile
pictures a lot, so I added #small_user_picture and #big_user_picture as
helper to users_helper.