Add #user_links to users_helper
This method takes the user as an argument and generates a link to the
user's profile with the user's profile picture and name.

I have updated all the profile picture + name pairs to use this helper
and made the profile pictures next to the posts also link to the user's
profile.