Add sticky header + labels to post form
Added a sticky header for convenience, modified some titles from h3 to h2
and added labels to the post form.

The sticky header was added with inline css. I know this isn't best practice,
but it's the most convenient for what this project is at the moment. (a
way to practice my backend Rails knowledge)