Add a new stimulus controller for post#new form
All it does is to clear the input of the form's fields after submission,
using the requestAnimationFrame javascript function to ensure that the
fields only get cleared after the form submission and right before the
DOM gets updated with the new post.

The 2nd part is not as important, but clearing them before the form submission
would mean that the form won't submit.