Add ability to unlike a liked post
Added a destroy action to the controller that handles this and turbo_stream
for both the create and destroy actions that update the like count and
change between the create and destroy forms in the DOM.

Two partials were made for this, _form and _unlike_form, which are very
similar, the only difference being the method, model and url. I'll have
to look into a way to combine them so I don't repeat myself there, if one
exists.

I've also changed the way posts are rendered on the user page so that they
use the _post partial, which is something I should have done when that
partial was created but I forgot to.