Add the ability to comment on a post
Single commit because this was fairly straight forward the second time
around. I created the model, established the relationships, added a
controller with the create action using turbo_stream and made the required
form and create.turbo_stream.erb view to save a comment and prepend it
to the list of comments without a page reload.