Add stimulus to comment form and apending comments
The stimulus that empties the text field for the post form was added to
empty the text field of the comment form right before the comment is
rendered on the page.

Also changed the create.turbo_stream.erb of the comments to append comments,
not prepend them as that simply makes more sense.